MacOS installs

+++ install/apps/neovim.sh
@@ -1,13 +1,17 @@
 #!/bin/bash
 
-cd /tmp
-wget -O nvim.tar.gz "https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz"
-tar -xf nvim.tar.gz
-sudo install nvim-linux64/bin/nvim /usr/local/bin/nvim
-sudo cp -R nvim-linux64/lib /usr/local/
-sudo cp -R nvim-linux64/share /usr/local/
-rm -rf nvim-linux64 nvim.tar.gz
-cd -
+if [[ "$OS" == "Linux" ]]; then
+  cd /tmp
+  wget -O nvim.tar.gz "https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz"
+  tar -xf nvim.tar.gz
+  sudo install nvim-linux64/bin/nvim /usr/local/bin/nvim
+  sudo cp -R nvim-linux64/lib /usr/local/
+  sudo cp -R nvim-linux64/share /usr/local/
+  rm -rf nvim-linux64 nvim.tar.gz
+  cd -
+else
+  brew install neovim
+fi
 
 # Only attempt to set configuration if Neovim has never been run
 # if [ ! -d "$HOME/.config/nvim" ]; then
